增强功能 Squashed commit of the following:
commit 4e3d6206bee25393479ed98ab36d734bffdca055
Author: xzzpig <1146628972@qq.com>
Date:   Sun Oct 8 21:57:50 2017 +0800

    添加If,Exists,ExistInPath的JE

commit 59efd7d67b59756fc44e8ae1c7eed696e8cbd583
Author: xzzpig <1146628972@qq.com>
Date:   Tue Oct 3 19:53:05 2017 +0800

    优化search输出结果

commit 63157ea02a75546e5999e5a7b950d7dcebec1388
Author: xzzpig <1146628972@qq.com>
Date:   Tue Oct 3 19:42:10 2017 +0800

    完善install的错误提示

commit ddd3c4bbcd6db92069339819a176b62085ff4a84
Author: xzzpig <1146628972@qq.com>
Date:   Tue Oct 3 18:56:50 2017 +0800

    search修复alias名称无法搜索到的bug

commit f7a9326a422a9d16eacccc6847307d4bd8dc673d
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 23:22:29 2017 +0800

    添加update软件重名时的说明

commit 02c4d1efaccfb71c599733308bfdcb06b1f89549
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 23:16:05 2017 +0800

    添加aliasmap表，命令兼容别名

commit 11357b8bcaa262bcbdd470eb4272c8a4eec4c151
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 20:56:02 2017 +0800

    ShowCMD添加显示软件名称

commit 957baebf6c2ee0729011f1141f4e4336e7c7acb0
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 20:37:36 2017 +0800

    更改确认内容(压缩->解压)

commit 5d373aff3be470913e063283f4a5056addd7b58b
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 20:37:04 2017 +0800

    修复DownloadJE下载进度条不是10个*的bug

commit 443e676d545fa4e2542b00315573fcb2cb159aef
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 20:36:40 2017 +0800

    修复upgrade命令未解析Var的Bug

commit 263dfa8b68c09202b9da8a075548bb0b0da8b7c9
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 15:25:58 2017 +0800

    添加CheckMD5JE

commit 79e88e9d4a67278a4b8a0657a80b9d0e00f13d9f
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 15:13:07 2017 +0800

    修复needConfirm判断缺失Bug

commit 4044e1d2d02f3452db60c8fb2e7df4aad2bd520e
Author: xzzpig <1146628972@qq.com>
Date:   Mon Oct 2 15:12:24 2017 +0800

    安装时将align名改为原名

+++ Plugin/src/main/java/com/xzzpig/pigprogrammanager/json/DeleteJE.java
@@ -18,7 +18,9 @@ public class DeleteJE implements JsonExecutor {
         File file = new File(jsonObject.optString("file"));
         if (!file.exists())
             return true;
-        API.needConfirm("将删除文件:" + file.getAbsolutePath());
+        if (!API.needConfirm("将删除文件:" + file.getAbsolutePath())) {
+            throw new JsonExecuteException("DeleteFailedException", "ConfirmCanceled");
+        }
         if (!API.deleteDir(file))
             throw new JsonExecuteException("DeleteFailedException", "file(" + file.getAbsolutePath() + ") delete failed");
         return true;
